requirements.txt
pygame==2.5.2
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7
numpy==1.24.3
Pillow==9.5.0